Event and Property Object
Enforce policy on Event and Property object.
For Event object:
- Only event owner can generate or fire or trigger an event.
For Property object:
- Only property owner can setup setter and getter functions.
This is done by adding 'friend typename ...' on class template.
Note that 'friend typename ...' maybe a compiler specific feature
and/or may not comply with C++ standard. There is another way to
enforce those policy above but it requires too many changes in
the source code.